bot/pyproject.toml
[project]
name = "bot-factory-bot"
version = "0.1.0"
description = "Telegram bot for Bot Factory platform using aiogram"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.0.0",
    "python-dotenv>=1.0.0",
    "asgiref>=3.7.0",
    "Django>=5.0,<6.0",
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-unfold>=0.2.0",
    "django-cors-headers>=4.3.0",
    "django-filter>=23.3",
    "django-environ>=0.11.0",
    "psycopg2-binary>=2.9.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "python-dateutil>=2.8.2",
    "cryptography>=41.0.0",
    "Pillow>=10.0.0",
]

[tool.uv]
dev-dependencies = []

